app/requirements.txt
asgiref==3.5.2
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
